refactor(podcast.details.stats): refactor stats component
Linked to #33
